Removing circles by clicking on them - needs work
This gets rid of the circles when the user clicks on them. However, at the moment it doesn't get rid of the circles when they're in the bottom left of the delayBox so I need to work on that.

Pushing this to origin because the Undo button still works so that is a nice workaround until I get it working properly