Fixed circles not drawing and circles randomly disappearing issues
I had mixed up the Xs and Ys when i was searching for the circles to remove from the array.

Also added some code to debug the adding and removing of circles: it just prints the coordinates of the mouse event and the circle so you can check if they're correct.