Delay done!
Delay is fully implemented. Drawing circles in the delay box sets them as a delay in the buffer with a set gain. The circles can be drawn and removed as the sound is playing and the changes will be made in the current cycle. However, there are some audio glitches at the moment so I need to work on that. But other than that, everything functions as intended.